Because of amount of data toString is very wordy will cut down to essentials

+++ src/main/java/model/edu/cofc/cs656/models/Renter.java
@@ -4,7 +4,7 @@ public class Renter extends User{
 	
 	private boolean renterPaymentValidated;
 
-	public Renter(double userID, boolean isSignedUp, Profile userProfile) {
+	public Renter(int userID, boolean isSignedUp, Profile userProfile) {
 		super(userID, isSignedUp, userProfile);
 		this.renterPaymentValidated = false;
 		System.out.println("You've created a Renter");
